fix selectors and break up test suites
